# refactor io related part in table builder