--- a/fileutil/preallocate.go
+++ b/fileutil/preallocate.go
@@ -1,0 +1,14 @@
+package fileutil
+
+import "os"
+
+func Preallocate(f *os.File, size int64) error {
+	if size == 0 {
+		return nil
+	}
+	return preallocate(f, size)
+}
+
+func preallocateTrunc(f *os.File, size int64) error {
+	return f.Truncate(size)
+}