--- a/fileutil/preallocate_others.go
+++ b/fileutil/preallocate_others.go
@@ -1,0 +1,9 @@
+// +build !linux
+
+package fileutil
+
+import "os"
+
+func preallocate(f *os.File, size int64) error {
+	return preallocateTrunc(f, size)
+}