--- a/options/options.go
+++ b/options/options.go
@@ -32,4 +32,12 @@ const (
 type TableBuilderOptions struct {
 	EnableHashIndex bool
 	HashUtilRatio   float32
+	WriteBufferSize int
+	BytesPerSync    int
+	BytesPerSecond  int
+}
+
+type ValueLogWriterOptions struct {
+	WriteBufferSize int
+	BytesPerSync    int
 }