--- a/table/builder.go
+++ b/table/builder.go
@@ -18,7 +18,10 @@ package table
 
 import (
 	"encoding/binary"
+	"github.com/coocood/badger/fileutil"
 	"github.com/coocood/badger/options"
+	"golang.org/x/time/rate"
+	"os"
 	"reflect"
 	"unsafe"
 
@@ -51,8 +54,9 @@ const headerSize = 4
 type Builder struct {
 	counter int // Number of keys written for the current block.
 
-	// Typically tens or hundreds of meg. This is for one single file.
-	buf []byte
+	w          *fileutil.BufferedFileWriter
+	buf        []byte
+	writtenLen int
 
 	baseKeysBuf     []byte
 	baseKeysEndOffs []uint32
@@ -74,10 +78,11 @@ type Builder struct {
 
 // NewTableBuilder makes a new TableBuilder.
 // The initCap is used to avoid memory reallocation.
-func NewTableBuilder(initCap int64, opt options.TableBuilderOptions) *Builder {
+func NewTableBuilder(f *os.File, limiter *rate.Limiter, opt options.TableBuilderOptions) *Builder {
 	assumeKeyNum := 256 * 1024
 	return &Builder{
-		buf:         make([]byte, 0, initCap),
+		w:           fileutil.NewBufferedFileWriter(f, opt.WriteBufferSize, opt.BytesPerSync, limiter),
+		buf:         make([]byte, 0, 4*1024),
 		baseKeysBuf: make([]byte, 0, assumeKeyNum/restartInterval),
 		// assume a large enough num of keys to init bloom filter.
 		bloomFilter:      bbloom.New(float64(assumeKeyNum), 0.01),
@@ -86,11 +91,35 @@ func NewTableBuilder(initCap int64, opt options.TableBuilderOptions) *Builder {
 	}
 }
 
+func (b *Builder) Reset(f *os.File) {
+	b.resetBuffers()
+	b.w.Reset(f)
+}
+
+func (b *Builder) ResetWithLimiter(f *os.File, limiter *rate.Limiter) {
+	b.resetBuffers()
+	b.w.ResetWithLimiter(f, limiter)
+}
+
+func (b *Builder) resetBuffers() {
+	b.counter = 0
+	b.buf = b.buf[:0]
+	b.writtenLen = 0
+	b.baseKeysBuf = b.baseKeysBuf[:0]
+	b.baseKeysEndOffs = b.baseKeysEndOffs[:0]
+	b.blockBaseKey = b.blockBaseKey[:0]
+	b.blockBaseOffset = 0
+	b.blockEndOffsets = b.blockEndOffsets[:0]
+	b.entryEndOffsets = b.entryEndOffsets[:0]
+	b.bloomFilter.Clear()
+	b.hashIndexBuilder.reset()
+}
+
 // Close closes the TableBuilder.
 func (b *Builder) Close() {}
 
 // Empty returns whether it's empty.
-func (b *Builder) Empty() bool { return len(b.buf) == 0 }
+func (b *Builder) Empty() bool { return b.writtenLen+len(b.buf) == 0 }
 
 // keyDiff returns a suffix of newKey that is different from b.blockBaseKey.
 func (b Builder) keyDiff(newKey []byte) []byte {
@@ -130,14 +159,14 @@ func (b *Builder) addHelper(key []byte, v y.ValueStruct) {
 	b.buf = append(b.buf, h.Encode()...)
 	b.buf = append(b.buf, diffKey...) // We only need to store the key difference.
 	b.buf = v.EncodeTo(b.buf)
-	b.entryEndOffsets = append(b.entryEndOffsets, uint32(len(b.buf))-b.blockBaseOffset)
+	b.entryEndOffsets = append(b.entryEndOffsets, uint32(b.writtenLen+len(b.buf))-b.blockBaseOffset)
 	b.counter++ // Increment number of keys added for this current block.
 }
 
-func (b *Builder) finishBlock() {
+func (b *Builder) finishBlock() error {
 	b.buf = append(b.buf, u32SliceToBytes(b.entryEndOffsets)...)
 	b.buf = append(b.buf, u32ToBytes(uint32(len(b.entryEndOffsets)))...)
-	b.blockEndOffsets = append(b.blockEndOffsets, uint32(len(b.buf)))
+	b.blockEndOffsets = append(b.blockEndOffsets, uint32(b.writtenLen+len(b.buf)))
 
 	// Add base key.
 	b.baseKeysBuf = append(b.baseKeysBuf, b.blockBaseKey...)
@@ -147,14 +176,22 @@ func (b *Builder) finishBlock() {
 	b.entryEndOffsets = b.entryEndOffsets[:0]
 	b.counter = 0
 	b.blockBaseKey = b.blockBaseKey[:0]
-	b.blockBaseOffset = uint32(len(b.buf))
+	b.blockBaseOffset = uint32(b.writtenLen + len(b.buf))
+	b.writtenLen += len(b.buf)
+	if err := b.w.Append(b.buf); err != nil {
+		return err
+	}
+	b.buf = b.buf[:0]
+	return nil
 }
 
 // Add adds a key-value pair to the block.
 // If doNotRestart is true, we will not restart even if b.counter >= restartInterval.
 func (b *Builder) Add(key []byte, value y.ValueStruct) error {
 	if b.counter >= restartInterval {
-		b.finishBlock()
+		if err := b.finishBlock(); err != nil {
+			return err
+		}
 	}
 	b.addHelper(key, value)
 	return nil // Currently, there is no meaningful error.
@@ -167,15 +204,15 @@ func (b *Builder) Add(key []byte, value y.ValueStruct) error {
 
 // ReachedCapacity returns true if we... roughly (?) reached capacity?
 func (b *Builder) ReachedCapacity(capacity int64) bool {
-	estimateSz := len(b.buf) +
+	estimateSz := b.writtenLen + len(b.buf) +
 		4*len(b.blockEndOffsets) +
 		len(b.baseKeysBuf) +
 		4*len(b.baseKeysEndOffs)
 	return int64(estimateSz) > capacity
 }
 
-// Finish finishes the table by appending the index.
-func (b *Builder) Finish() []byte {
+// Flush finishes the table by appending the index.
+func (b *Builder) Finish() error {
 	b.finishBlock() // This will never start a new block.
 	b.buf = append(b.buf, u32SliceToBytes(b.blockEndOffsets)...)
 	b.buf = append(b.buf, b.baseKeysBuf...)
@@ -193,7 +230,7 @@ func (b *Builder) Finish() []byte {
 		b.buf = append(b.buf, u32ToBytes(0)...)
 	}
 
-	return b.buf
+	return b.w.FlushWithData(b.buf, true)
 }
 
 func u32ToBytes(v uint32) []byte {