## Motivation
There are some shortcomings in badger’s IO related part.
* Use `O_DSYNC` instead of `fdatasync`, besides some platform define `O_DSYNC` but doesn’t respect to it (fortunately Linux 3.2.0+ works well), this flag also restrict the way we write data to file. Because each `write` call will sync data, so we must keep as much as possible data in buffer instead of let it in system cache.
* Unfortunately, due to badger doesn’t preallocate file with desired size, each `write` will extend file size. In this case, `fdatasync (O_DSYNC)` have to update file’s metadata.
* To control the IO spike, currently we break table flush to small chunks and write them chuck by chunk. But according to previous reason, it cause a lots of metadata update. As final result, the flushing process takes a longer time to finish, occupying online queries’ IO bandwidth for a longer time.

## What Is Changed And How It Works
* Introduce a `buffered file writer` in `vlog writer` and `table builder` . Instead of use `O_DSYNC` when open file, call `fdatasync` at the end of file writing. After this change, `table builder` doesn’t need to allocate a large buffer for each table and start a goroutine to write. It can reuse the `builder` instance, and call `write` while iteration. 
* Preallocating `value log` file. After this change, `fdatasync` can really reduce IO of update metadata.
* To control the IO spike:
	* Using `sync_file_range` to hint OS to flush some dirty pages . Linux has a useful sync function, `sync_file_range`, it can flush specific dirty pages to disk and **never** update metadata. `buffered file writer` use it internally, controlled by `BytesPerSync`. With this change, we can control the total number of dirty pages in cache, avoid a burst IO when call `fdatasync`.
	* Introduce `rate limiter` in `table builder`. `buffered file writer` also support `Token Buckets` rate limiter. We can config the max write speed during compaction, reserve enough IO bandwidth for online queries.
	* Combine the two method, we can control the IO of compaction more reliable than current implementation.
* Introduce a bunch of options to config this new features.

## Benchmark Result Of Sysbench
### Config
```
nodes:
192.168.199.105: pd，sysbench，TiDB
192.168.199.104, 192.168.199.180, 192.168.199.100: TiDB
192.168.199.106: unistore

sysbench:
tables=1 
table-size=10000000
512 connections per TiDB

TiDB:
[prepared-plan-cache]
enabled = true
[txn-local-latches]
enabled = false 

unistore:
table-loading-mode = memory-map
num-mem-tables = 3
num-level-zero-tables = 3
num-index-dbs = 2
num-row-dbs = 2
hash-ratio = 0.75
```

Config of `this branch`
```
TableBuilderOptions.BytesPerSync = 2MB
TableBuilderOptions.WriteBufferSize = 1MB
TableBuilderOptions.BytesPerSecond = 20MB
ValueLogWriteOptions.WriteBufferSize = 8MB
ValueLogWriteOptions.BytesPerSync = 0 # disabled
```
Others config keeps the same with `master`

### update_index
**master**
```
SQL statistics:
    queries performed:
        read:                            0
        write:                           4896925
        other:                           0
        total:                           4896925
    transactions:                        4896925 (8094.49 per sec.)
    queries:                             4896925 (8094.49 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          604.9675s
    total number of events:              4896925

Latency (ms):
         min:                                   61.80
         avg:                                  250.98
         max:                                10358.28
         95th percentile:                      376.49
         sum:                           1229014495.25

Threads fairness:
    events (avg/stddev):           2391.0767/29.39
    execution time (avg/stddev):   600.1047/0.25
```

**this branch**
```
SQL statistics:
    queries performed:
        read:                            0
        write:                           5213163
        other:                           0
        total:                           5213163
    transactions:                        5213163 (8629.19 per sec.)
    queries:                             5213163 (8629.19 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          604.1294s
    total number of events:              5213163

Latency (ms):
         min:                                   60.34
         avg:                                  235.76
         max:                                10831.32
         95th percentile:                      404.61
         sum:                           1229030647.02

Threads fairness:
    events (avg/stddev):           2545.4897/32.64
    execution time (avg/stddev):   600.1126/0.26
```
The `QPS` and `95% late` during table compaction, keeps nearly the same between these two branch.
### update_non_index
**master**
```
SQL statistics:
    queries performed:
        read:                            0
        write:                           8713622
        other:                           0
        total:                           8713622
    transactions:                        8713622 (14516.39 per sec.)
    queries:                             8713622 (14516.39 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          600.2594s
    total number of events:              8713622

Latency (ms):
         min:                                   58.72
         avg:                                  141.03
         max:                                 1461.83
         95th percentile:                      257.95
         sum:                           1228860921.23

Threads fairness:
    events (avg/stddev):           4254.6982/277.47
    execution time (avg/stddev):   600.0297/0.04
```

**this branch**
```
SQL statistics:
    queries performed:
        read:                            0
        write:                           10407699
        other:                           0
        total:                           10407699
    transactions:                        10407699 (17338.18 per sec.)
    queries:                             10407699 (17338.18 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          600.2745s
    total number of events:              10407699

Latency (ms):
         min:                                   51.78
         avg:                                  118.08
         max:                                 1469.13
         95th percentile:                      200.47
         sum:                           1228919206.10

Threads fairness:
    events (avg/stddev):           5081.8843/388.01
    execution time (avg/stddev):   600.0582/0.05
```
The `QPS` during compaction of `this branch` is much higher than `master` (around 12k vs around 8k)