@bobotu 2018-11-12
Sorry for forget some functions' doc and comment, update in later commit.☺️